+++ backend2/src/main/java/com/cmoney/backend2/virtualtrading2/service/api/tseotc/deletedelagate/DeleteDelegateResponseBody.kt
@@ -1,0 +1,14 @@
+package com.cmoney.backend2.virtualtrading2.service.api.tseotc.deletedelagate
+
+import com.google.gson.annotations.SerializedName
+
+/**
+ * 刪除委託單回應
+ *
+ * @property delegateId 委託單號
+ *
+ */
+data class DeleteDelegateResponseBody(
+    @SerializedName("OrderNo")
+    val delegateId: Long?
+)