+++ CHANGELOG.md
@@ -1,8 +1,40 @@
-#  (2023-02-07)
+# [5.50.0](http://192.168.10.147/CG_Mobile/CG_Module_Android/Backend2/backend2/compare/5.49.0...5.50.0)
 
 
+### Features
 
-## [5.48.1](http://192.168.10.147/CG_Mobile/CG_Module_Android/Backend2/backend2/compare/5.48.0...5.48.1) (2023-02-07)
+- 新增釘選社團看板文章
+- 新增取消釘選社團看板文章
+- 新增取得社團看板置頂文章列表
+- 新增取得社團看板置頂文章列表
+- 將toListOfSomething()棄用，改用toListOfType()
+
+
+# [5.49.0](http://192.168.10.147/CG_Mobile/CG_Module_Android/Backend2/backend2/compare/5.48.1...5.49.0)
+
+
+### Features
+
+- 新增全域的Backend2管理者(目前能控制虛擬下單Domain)
+- 新增可以變更Domain和特定Path的物件，詳看VirtualTradingV2資料夾
+
+
+#### 虛擬下單
+
+- 新增虛擬下單V2-建立帳號
+- 新增虛擬下單V2-建立上市上櫃委託單
+- 新增虛擬下單V2-刪除上市上櫃委託單
+- 新增虛擬下單V2-取得全部帳號資訊
+- 新增虛擬下單V2-取得特定帳號資訊
+- 新增虛擬下單V2-取得帳號報酬率
+- 新增虛擬下單V2-取得上市上櫃委託單
+- 新增虛擬下單V2-取得上市上櫃委託單細節
+- 新增虛擬下單V2-取得上市上櫃所有的成交單
+- 新增虛擬下單V2-取得上市上櫃的成交單細節
+- 新增虛擬下單V2-取得上市上櫃的庫存
+
+
+## [5.48.1](http://192.168.10.147/CG_Mobile/CG_Module_Android/Backend2/backend2/compare/5.48.0...5.48.1)
 
 
 ### Changes